Clean up ShadowMap implementation
* Additionally gave the CircadianCycle a start/stop and speed slider in
the UI.

+++ Sprocket/Systems/PlayerMovement.h
@@ -7,7 +7,7 @@ namespace Sprocket {
 
 class PlayerMovement : public EntitySystem
 {
-    float d_deltaTime;
+    double d_deltaTime;
 
     bool d_enabled = true;
